chore: improved code structure for new processors
refs: #37
