+++ src/processors/process-auto-location.func.ts
@@ -1,0 +1,3 @@
+export const processAutoLocationCodeBlock = (source: string) => {
+	console.log("I am the processAutoLocationCodeBlock function");
+};
